Updated schemas and updated database operations

+++ routes/notesRoute.js
@@ -1,11 +1,8 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const mongoose = require('mongoose');
 const cors = require('./cors');
 const authenticate = require('../authenticate');
 const Notes = require('../models/notes');
-const User = require('../models/user');
-const { router } = require('../app');
 
 const notesRouter = express.Router();
 notesRouter.use(bodyParser.json());
@@ -13,8 +10,9 @@ notesRouter.use(bodyParser.json());
 
 notesRouter.route('/')
 .options(cors.corsWithOptions, (req, res) => { res.sendStatus(200); })
-.get(cors.cors, (req,res,next) => {
-    Notes.find({})
+.get(cors.cors, authenticate.verifyUser,(req,res,next) => {
+    Notes.find({author: req.user._id})
+    .populate('author')
     .then((notes) => {
         res.statusCode = 200;
         res.setHeader('Content-Type', 'application/json');
@@ -24,33 +22,63 @@ notesRouter.route('/')
 })
 
 .post(cors.cors, authenticate.verifyUser, (req,res,next) => {
-    Notes.create(req.body)
-    .then((note) => {
-        res.statusCode = 200;
+    if (req.body != null){
+        req.body.author = req.user._id;
+        Notes.create(req.body)
+        .then((note) => {
+            Notes.findById(note._id)
+            .populate('author')
+            .then((note) => {
+                res.statusCode = 200;
+                res.setHeader('Content-Type', 'application/json');
+                res.json(note);
+            })
+
+        }, err => next(err))
+        .catch(err => next(err));
+    }
+    else {
+        res.statusCode = 404;
         res.setHeader('Content-Type', 'application/json');
-        res.json(note);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+        res.json({success: false, message: 'No data was found in the request body'});
+    }
 })
 
 .delete(cors.cors, authenticate.verifyUser, (req,res,next) => {
-    Notes.remove({})
-    .then((response) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(response);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    res.statusCode = 403;
+    res.setHeader('Content-Type', 'application/json');
+    res.json({success: false, message: 'UnAuthorised access denied' });
+    // Notes.remove({})
+    // .then((response) => {
+    //     res.statusCode = 200;
+    //     res.setHeader('Content-Type', 'application/json');
+    //     res.json(response);
+    // }, (err) => next(err))
+    // .catch((err) => next(err));
 });
 
 notesRouter.route('/:notesId')
-.get(cors.cors, (req,res,next) => {
+.get(cors.cors, authenticate.verifyUser, (req,res,next) => {
     Notes.findById(req.params.notesId)
+    .populate('author')
     .then((note) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(note);
-    }, (err) => next(err))
+        if (note.author.equals(req.user._id)){
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(note);
+        }
+        else {
+            res.statusCode = 403;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'You are not authorized to view this note' });
+        }
+    }, (err) => {
+        if (err){
+            res.statusCode = 404;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'Note with ID '+ req.params.notesId + ' was not found!' });
+        }
+    })
     .catch((err) => next(err));
 })
 
@@ -60,23 +88,65 @@ notesRouter.route('/:notesId')
 })
 
 .put(cors.cors, authenticate.verifyUser, (req,res,next) => {
-    Notes.findByIdAndUpdate(req.params.notesId, {
-        $set: req.body
-    }, {new: true})
+    Notes.findById(req.params.notesId)
     .then((note) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(note);
+        if (note != null){
+            if (note.author.equals(req.user._id)){
+                req.body.author = req.user._id;
+                Notes.findByIdAndUpdate(req.params.notesId, {
+                    $set: req.body
+                }, { new: true })
+                .then((note) => {
+                    Notes.findById(note._id)
+                    .populate('author')
+                    .then((note) => {
+                        res.statusCode = 200;
+                        res.setHeader('Content-Type', 'application/json');
+                        res.json(note); 
+                    })               
+                }, (err) => next(err));
+            }
+            else {
+                res.statusCode = 403;
+                res.setHeader('Content-Type', 'application/json');
+                res.json({success: false, message: 'You are not authorized to modify this note' });
+            }
+        }
+        else {
+            res.statusCode = 404;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'Note with ID '+ req.params.notesId + ' was not found!' });
+        }
     }, (err) => next(err))
     .catch((err) => next(err));
 })
 
 .delete(cors.cors, authenticate.verifyUser, (req,res,next) => {
-    Notes.findByIdAndRemove(req.params.notesId)
-    .then((response) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(response);
+    Notes.findById(req.params.notesId)
+    .then((note) => {
+        if (note != null){
+            if (note.author.equals(req.user._id)){
+                Notes.findByIdAndRemove(req.params.notesId)
+                .then((note) => {
+                    Notes.findById(note._id)
+                    .then((note) => {
+                        res.statusCode = 200;
+                        res.setHeader('Content-Type', 'application/json');
+                        res.json(note); 
+                    })               
+                }, (err) => next(err));
+            }
+            else {
+                res.statusCode = 403;
+                res.setHeader('Content-Type', 'application/json');
+                res.json({success: false, message: 'You are not authorized to delete this note' });
+            }
+        }
+        else {
+            res.statusCode = 404;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'Note with ID '+ req.params.notesId + ' was not found!' });
+        }
     }, (err) => next(err))
     .catch((err) => next(err));
 });
